feat: 添加了应用程序图标、托盘图标的支持
Resolved #8
