+++ .gitignore
@@ -1,6 +1,5 @@
 /target
-/assets
+/assets/images
 /.idea
 /baks
 .links.json
-logo.*
