+++ Cargo.toml
@@ -26,3 +26,7 @@ strsim = "0.11.1"
 
 tokio = { version = "1.45.0", features = ["full"] }
 single-instance = "0.3.3"
+image = "0.25.6"
+
+[build-dependencies]
+winres = "0.1"
