+++ src/main.rs
@@ -2,12 +2,12 @@
 
 mod my_structs;
 mod pages;
+mod resources;
 mod window;
 
 use std::sync::{Arc, Mutex};  // Arc = 原子引用计数(Atomically Reference Counted)，一种线程安全的智能指针，允许在多个线程间共享所有权
 use egui_winit::winit;
 use rdev::{listen, EventType, Key};
-use winit::platform::windows::IconExtWindows;
 use std::time::{Duration, Instant};
 use trayicon;
 use single_instance::SingleInstance;
@@ -17,7 +17,7 @@ use my_structs::MyApp;
 
 
 pub const WINDOW_SIZE: (f32, f32) = (800.0, 500.0);
-pub const PROGRAM_VERSION: &str = "v0.1.3-alpha.1";
+pub const PROGRAM_VERSION: &str = "v0.1.3-alpha.02";
 pub const CONFIG_FILE_VERSION: u32 = 4;
 pub const CONFIG_FILE_NAME: &str = ".links.json";
 pub const DOUBLE_ALT_COOLDOWN: u64 = 500;
@@ -51,8 +51,6 @@ fn main() {
 
     let called = Arc::new(Mutex::new(true));
     let called_clone = called.clone();
-
-    
 
     rt.spawn(async move {
         // loop {
@@ -127,7 +125,12 @@ fn main() {
             height: WINDOW_SIZE.1,
         })
         .with_title("BaroBoard 工具箱") // 参见 https://github.com/emilk/egui/pull/2279
-        // .with_window_icon(Some(winit::window::Icon::from_rgba(rgba, width, height)))
+        .with_window_icon({
+            let rgba = image::load_from_memory(resources::LOGO_ICO).unwrap().to_rgba8();
+            let (width, height) = rgba.dimensions();
+            let rgba_data = rgba.into_raw();
+            Some(winit::window::Icon::from_rgba(rgba_data, width, height).unwrap())
+        })
         // .with_visible(false)
         ;
 
@@ -138,7 +141,7 @@ fn main() {
     .sender(move |e: &event::UserEvent| {
         let _ = proxy_clone_tray.send_event(e.clone());
     })
-    .icon_from_buffer(include_bytes!("../logo.ico"))
+    .icon_from_buffer(resources::LOGO_ICO)
     .tooltip("BaroBoard 工具箱")
 
     .on_click(event::UserEvent::LeftClickTrayIcon)
@@ -168,7 +171,7 @@ fn main() {
             // 安装图片加载器，允许egui加载和显示图片
             egui_extras::install_image_loaders(egui_ctx);
             // 设置自定义字体，支持中文显示
-            setup_custom_fonts(egui_ctx);
+            resources::setup_custom_fonts(egui_ctx);
 
             Box::new(MyApp::new(
                 called.clone(),
@@ -183,80 +186,4 @@ fn main() {
     // 启动事件循环，这通常是阻塞的，会一直运行直到应用程序关闭
     // 事件循环会不断处理输入事件、UI更新和渲染，这是GUI应用程序的主要执行模式
     event_loop.run_app(&mut app).expect("failed to run app");
-}
-
-
-//自定义字体
-fn setup_custom_fonts(ctx: &egui::Context) {
-    // 创建一个默认的字体定义对象
-    let mut fonts = egui::FontDefinitions::default();
-
-    // 根据不同操作系统选择不同的字体路径
-    let font_path = if cfg!(target_os = "windows") {
-        // Windows系统下微软雅黑的默认位置
-        std::path::Path::new("C:/Windows/Fonts/msyh.ttc")
-    } else if cfg!(target_os = "linux") {
-        // Linux系统下常见的中文字体路径，尝试几个常见位置
-        let possible_paths = [
-            "/usr/share/fonts/noto/NotoSansCJK-Regular.ttc",         // Noto Sans CJK
-            "/usr/share/fonts/wenquanyi/wqy-microhei.ttc",           // 文泉驿微米黑
-            "/usr/share/fonts/wenquanyi/wqy-zenhei.ttc",             // 文泉驿正黑
-            "/usr/share/fonts/truetype/droid/DroidSansFallbackFull.ttf", // Droid Sans
-        ];
-        
-        // 查找第一个存在的字体文件
-        let mut found_path = std::path::Path::new("/usr/share/fonts"); // 默认路径
-        for path in possible_paths {
-            let full_path = std::path::Path::new(path);
-            if full_path.exists() {
-                found_path = full_path;
-                break;
-            }
-        }
-        found_path
-    } else if cfg!(target_os = "macos") {
-        // macOS系统下常见的中文字体
-        std::path::Path::new("/System/Library/Fonts/PingFang.ttc")  // 苹方字体
-    } else {
-        // 其他操作系统使用一个不太可能存在的路径，将回退到默认字体
-        std::path::Path::new("/non-existent-path")
-    };
-    
-    if font_path.exists() {
-        // 如果找到字体文件，从文件读取
-        match std::fs::read(font_path) {
-            Ok(font_data) => {
-                println!("已加载系统字体: {:?}", font_path);
-                fonts.font_data.insert(
-                    "my_font".to_owned(),
-                    // 这里也使用Arc共享字体数据，但这里的Arc主要用于避免数据复制，而非线程安全
-                    // 在egui中，Arc用于智能地共享大型资源(如字体)，减少内存使用
-                    Arc::new(egui::FontData::from_owned(font_data)),
-                );
-                
-                // 将字体添加到 Proportional 字体族的第一个位置
-                fonts
-                    .families
-                    .entry(egui::FontFamily::Proportional)
-                    .or_default()
-                    .insert(0, "my_font".to_owned());
-
-                // 将字体添加到 Monospace 字体族的末尾
-                fonts
-                    .families
-                    .entry(egui::FontFamily::Monospace)
-                    .or_default()
-                    .push("my_font".to_owned());
-            },
-            Err(err) => {
-                eprintln!("无法加载系统字体 {:?}: {}", font_path, err);
-                // 加载失败时使用默认字体
-            }
-        }
-    } else {
-        eprintln!("未找到系统字体 {:?}，将使用默认字体", font_path);
-    }
-
-    // 将字体设置应用到 egui 上下文
-    ctx.set_fonts(fonts);
 }
